Start the verification of plastic strain computation
Until now, a significant approximation was made in the plstic model,
considering that the 'strains' defined in the Piecewise Constant Strain
discrete Cosserat could be used in the same way as components of a
Green-Lagrange strain tensor. This is not true, as the 'strains' in the
Cosserat model are defined in a specific formalism.
This commit initialises the implementation of a correspondance between
the two types of strain (Cosserat vs Green-Lagrange classic formalism).
At first, we compute an estimation of the Green-Lagrange tensor based on
a finite differences computation at various points inside the Cosserat
medium. The 3D position of the tracked points, in the global frame, is
obtained using the Cosserat mapping (similarly as what is done for the
mapped frames)